# API/src/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Fitness API",
    "description": "A comprehensive fitness tracking and coaching API that allows coaches to manage trainees, create workout and nutrition plans, and track progress.",
    "contact": {
      "name": "API Support",
      "email": "support@fitness-api.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication"
      }
    },
    "schemas": {
...
      "SuccessMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "CreateWorkoutPlan": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Name of the workout plan",
            "example": "Push Pull Legs"
          },
          "description": {
            "type": "string",
            "description": "Description of the workout plan",
            "example": "A 6-day split focusing on push, pull, and leg movements"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "description": "Difficulty level of the workout",
            "example": "intermediate"
          },
          "estimatedCalories": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Estimated calories burned per session",
            "example": 350
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateWorkoutPlan": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Updated name of the workout plan"
          },
          "description": {
            "type": "string",
            "description": "Updated description of the workout plan"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "description": "Updated difficulty level"
          },
          "estimatedCalories": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Updated estimated calories burned"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the workout plan is active",
            "example": true
          }
        }
      },
      "AssignWorkoutPlan": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "minLength": 1,
            "description": "ID of the user to assign the plan to",
            "example": "user456"
          },
          "startDate": {
            "type": "string",
            "description": "Start date for the workout plan",
            "example": "2025-01-01"
          },
          "endDate": {
            "type": "string",
            "description": "End date for the workout plan",
            "example": "2025-06-01"
          }
        },
        "required": [
          "userId",
          "startDate"
        ]
      },
...
      "WorkoutPlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Workout plan ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Plan name",
            "example": "Push Pull Legs"
          },
          "description": {
            "type": "string",
            "description": "Plan description"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ]
          },
          "estimatedCalories": {
            "type": "number",
            "description": "Estimated calories per session"
          },
          "createdBy": {
            "type": "string",
            "description": "Creator user ID"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether plan is active"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation date"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "name",
          "createdBy",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "RecordUserStats": {
        "type": "object",
        "properties": {
          "weight": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Body weight in kg",
            "example": 75.5
          },
          "height": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Height in cm",
            "example": 175
          },
          "bodyFat": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "description": "Body fat percentage",
            "example": 15.2
          },
          "muscleMass": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Muscle mass in kg",
            "example": 35.8
          },
          "chest": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Chest measurement in cm",
            "example": 102
          },
          "waist": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Waist measurement in cm",
            "example": 85
          },
          "hips": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Hip measurement in cm",
            "example": 95
          },
          "arms": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Arm measurement in cm",
            "example": 38
          },
          "thighs": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Thigh measurement in cm",
            "example": 58
          },
          "notes": {
            "type": "string",
            "description": "Additional notes about measurements",
            "example": "Measured in the morning"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID",
            "example": "user123"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email",
            "example": "john.doe@example.com"
          },
          "name": {
            "type": "string",
            "description": "User full name",
            "example": "John Doe"
          },
          "role": {
            "type": "string",
            "enum": [
              "coach",
              "trainee"
            ],
            "description": "User role",
            "example": "trainee"
          },
          "createdAt": {
            "type": "string",
            "description": "Account creation date",
            "example": "2025-01-01T00:00:00Z"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "createdAt"
        ]
      },
...
      "WorkoutPlanAssignment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "userId": {
            "type": "string"
          },
          "workoutPlanId": {
            "type": "number"
          },
          "assignedBy": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "assignedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "workoutPlanId",
          "assignedBy",
          "startDate",
          "isActive",
          "assignedAt"
        ]
      },
      "UserStats": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "userId": {
            "type": "string"
          },
          "weight": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "bodyFat": {
            "type": "number"
          },
          "muscleMass": {
            "type": "number"
          },
          "chest": {
            "type": "number"
          },
          "waist": {
            "type": "number"
          },
          "hips": {
            "type": "number"
          },
          "arms": {
            "type": "number"
          },
          "thighs": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          },
          "recordedAt": {
            "type": "string"
          },
          "recordedBy": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "recordedAt",
          "recordedBy"
        ]
      },
      "UserStatsResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "userId": {
            "type": "string"
          },
          "weight": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "bodyFat": {
            "type": "number"
          },
          "muscleMass": {
            "type": "number"
          },
          "chest": {
            "type": "number"
          },
          "waist": {
            "type": "number"
          },
          "hips": {
            "type": "number"
          },
          "arms": {
            "type": "number"
          },
          "thighs": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          },
          "recordedAt": {
            "type": "string"
          },
          "recordedBy": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "recordedAt",
          "recordedBy"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/workouts": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Get all workout plans",
        "description": "Retrieve workout plans, optionally filtered by current user",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Filter by current user resources",
              "example": "true"
            },
            "required": false,
            "description": "Filter by current user resources",
            "name": "my",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of workout plans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutPlan"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workouts"
        ],
        "summary": "Create workout plan",
        "description": "Create a new workout plan",
        "requestBody": {
          "description": "Workout plan data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutPlan"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workout plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutPlan"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/api/workouts/{id}": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Get workout plan by ID",
        "description": "Retrieve a specific workout plan with details",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "Unique identifier",
              "example": "123"
            },
            "required": true,
            "description": "Unique identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Workout plan details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutPlan"
                }
              }
            }
          },
          "404": {
            "description": "Workout plan not found"
          }
        }
      },
      "put": {
        "tags": [
          "Workouts"
        ],
        "summary": "Update workout plan",
        "description": "Update an existing workout plan (coach only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "Unique identifier",
              "example": "123"
            },
            "required": true,
            "description": "Unique identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "description": "Updated workout plan data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkoutPlan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workout plan updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutPlan"
                }
              }
            }
          },
          "404": {
            "description": "Workout plan not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Workouts"
        ],
        "summary": "Delete workout plan",
        "description": "Delete a workout plan (coach only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "Unique identifier",
              "example": "123"
            },
            "required": true,
            "description": "Unique identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Workout plan deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                }
              }
            }
          },
          "404": {
            "description": "Workout plan not found"
          }
        }
      }
    },
    "/api/workouts/{id}/assign": {
      "post": {
        "tags": [
          "Workouts"
        ],
        "summary": "Assign workout plan to user",
        "description": "Assign a workout plan to a trainee (coach only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "Unique identifier",
              "example": "123"
            },
            "required": true,
            "description": "Unique identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "description": "Assignment data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignWorkoutPlan"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workout plan assigned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutPlanAssignment"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/stats": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user body stats",
        "description": "Get body measurement statistics for the current user",
        "responses": {
          "200": {
            "description": "User body stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserStats"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Record user body stats",
        "description": "Record new body measurements for the current user",
        "requestBody": {
          "description": "Body measurement data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordUserStats"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Body stats recorded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/search": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Search users",
        "description": "Search for users by name or email, optionally filtered by role",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Search term for finding users",
              "example": "john"
            },
            "required": true,
            "description": "Search term for finding users",
            "name": "query",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "coach",
                "trainee"
              ],
              "description": "Filter by user role",
              "example": "trainee"
            },
            "required": false,
            "description": "Filter by user role",
            "name": "role",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}